pdf2pps_backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
psycopg = {extras = ["binary"], version = "^3.2.6"}
pymupdf = "^1.25.4"
python-pptx = "^1.0.2"
ollama = "^0.4.7"
langchain = "^0.3.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
